preparacao pra  fazer o logging

+++ include/BPlusTree.hpp
@@ -1,17 +1,15 @@
 #ifndef BPLUS_TREE
 #define BPLUS_TREE
 
-#include <iostream>
 #include <vector>
 #include <fstream>
 #include <math.h>
-#include <queue>
 
 struct NoBplus {
     bool isFolha; //dita se um nó é folha ou não
-    std::vector<int> chaves;
+    std::vector<unsigned int> chaves;
     std::vector<NoBplus*> filhos; //somente para nós internos
-    std::vector<int> enderecos;  //somente para nós folhas
+    std::vector<unsigned long long> enderecos;  //somente para nós folhas
     NoBplus* proximo;            //somente para nós folhas
     NoBplus* pai;
 };
@@ -24,10 +22,10 @@ class BplusTree {
     public: 
         BplusTree(int ordem);
         NoBplus* criaNovoNo();
-        void insere(int chave, int endereco);
-        void insereNaFolha(NoBplus* no, int chave, int endereco);
-        void inserePai(NoBplus* noEsq, int chave, NoBplus* noDir);
-        NoBplus* buscaNo(int chave);
+        void insere(unsigned int chave, unsigned long long endereco);
+        void insereNaFolha(NoBplus* no, unsigned int chave, unsigned long long endereco);
+        void inserePai(NoBplus* noEsq, unsigned int chave, NoBplus* noDir);
+        NoBplus* buscaNo(unsigned int chave);
         void imprimeArvore();
 };
